+++ src/main/plugin-system/loadPlugins.ts
@@ -12,6 +12,7 @@ import coreSidebar from "../plugins/core-sidebar/index";
 import coreBookmarks from "../plugins/core-bookmarks/index";
 import coreContextMenu from "../plugins/core-context-menu/index";
 import coreWorkspace from "../plugins/core-workspace/index";
+import coreImageViewer from "../plugins/core-image-viewer/index";
 import thirdPartyBlockRef from "../plugins/third-party-block-ref/index";
 import thirdPartyPageMode from "../plugins/third-party-page-mode/index";
 import thirdPartyLinks from "../plugins/third-party-links/index";
@@ -28,6 +29,7 @@ const MAIN_PLUGINS: Record<string, MainPlugin> = {
   "core-bookmarks": coreBookmarks,
   "core-context-menu": coreContextMenu,
   "core-workspace": coreWorkspace,
+  "core-image-viewer": coreImageViewer,
   "third-party-block-ref": thirdPartyBlockRef,
   "third-party-page-mode": thirdPartyPageMode,
   "third-party-links": thirdPartyLinks,